start with simple refactor
