+++ guardrails/cli/__init__.py
@@ -1,0 +1,7 @@
+from guardrails.cli.cli import cli
+import guardrails.cli.validate
+import guardrails.cli.compile
+
+
+if __name__ == "__main__":
+    cli()
