+++ guardrails/cli/cli.py
@@ -1,0 +1,3 @@
+import typer
+
+cli = typer.Typer()
