+++ guardrails/cli/validate.py
@@ -1,39 +1,15 @@
 import json
 from typing import Dict, Union
 
-import typer
 
 from guardrails import Guard
-
-cli = typer.Typer()
-
-
-def compile_rail(rail: str, out: str) -> None:
-    """Compile guardrails from the guardrails.yml file."""
-    raise NotImplementedError("Currently compiling rail is not supported.")
-
+from guardrails.cli.cli import cli
 
 def validate_llm_output(rail: str, llm_output: str) -> Union[str, Dict, None]:
     """Validate guardrails.yml file."""
     guard = Guard.from_rail(rail)
     result = guard.parse(llm_output)
     return result.validated_output
-
-
-@cli.command()
-def compile(
-    rail: str = typer.Argument(
-        ..., help="Path to the rail spec.", exists=True, file_okay=True, dir_okay=False
-    ),
-    out: str = typer.Option(
-        default=".rail_output",
-        help="Path to the compiled output directory.",
-        file_okay=False,
-        dir_okay=True,
-    ),
-):
-    """Compile guardrails from a `rail` spec."""
-    print("Not supported yet. Use `validate` instead.")
 
 
 @cli.command()
@@ -59,7 +35,3 @@ def validate(
         f.write("\n")
 
     return result
-
-
-if __name__ == "__main__":
-    cli()
